Separate Versions
Different mod versions use different gradle versions, so we'll separate the versions to different folders
