+++ 1.21/src/main/resources/antihit.mixins.json
