+++ 1.21/src/main/resources/assets/antihit/icon.png
