+++ 1.21/src/main/resources/assets/antihit/lang/en_us.json
@@ -1,0 +1,8 @@
+{
+  "key.categories.antihit": "AntiHit",
+  "key.antihit.addfriend": "Add friend",
+  "key.antihit.removefriend": "Remove friend",
+  "key.antihit.togglefriendattack": "Toggle FriendAttack",
+  "key.antihit.removeallfriends": "Remove all Friends"
+
+}
