+++ 1.21/src/main/resources/fabric.mod.json
